shared/pyproject.toml
[project]
name = "agentic-layer-sdk"
version = "0.1.0"
readme = "README.md"
requires-python = ">=3.14,<3.15"
dependencies = [
    "loguru>=0.7.3",
    "pydantic>=2",
    "opentelemetry-exporter-otlp-proto-http",
    "opentelemetry-exporter-otlp-proto-grpc",
    "opentelemetry-instrumentation-starlette",
    "opentelemetry-instrumentation-httpx",
    "opentelemetry-instrumentation-logging",
    "starlette",
    "a2a-sdk",
]

[dependency-groups]
dev = [
    "mypy[reports]>=1.12.0,<2",
    "ruff>=0.11.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["agenticlayer"]
only-include = ["agenticlayer/shared"]

[tool.ruff]
line-length = 120

[tool.ruff.lint]
extend-select = ["Q", "N", "I001", "PT", "TID251", "DTZ"]

[tool.mypy]
strict = true
mypy_path = ["."]
namespace_packages = true
explicit_package_bases = true
